Refactor the first 5 containers into components
- AdminContainer
- AdvocateContainer
- AdvocateLoginContainer
- AdvocatesContainer
- AppContainer
