--- app/containers/AppContainer.jsx
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux'
-import App from '../components/App'
-
-function mapStateToProps (state) {
-  return {
-    auth: state.auth
-  }
-}
-
-export default connect(mapStateToProps)(App)
